Типы long и uint изменены на int

+++ Repositories/UserRepository.cs
@@ -5,11 +5,11 @@ namespace FactoryTracker.Repositories
 {
     public class UserRepository
     {
-        private readonly Dictionary<long, User> _users;
+        private readonly Dictionary<int, User> _users;
 
-        private long _userId = 0;
+        private int _userId = 0;
 
-        private long IncrementUserId()
+        private int IncrementUserId()
         {
             return _userId++;
         }
@@ -19,9 +19,9 @@ namespace FactoryTracker.Repositories
             _users.Add(IncrementUserId(), user);
         }
 
-        public User GetUser(long userId)
+        public User GetUser(int userId)
         {
-            KeyValuePair<long, User> foundItem = _users.FirstOrDefault(item => item.Key == userId);
+            KeyValuePair<int, User> foundItem = _users.FirstOrDefault(item => item.Key == userId);
             return foundItem.Value;
         }
     }
